backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
boto3>=1.29.0
botocore>=1.32.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
# ChromaDB is optional - install separately if needed: pip install chromadb
# For local dev, the vector store will work without it using mock responses

